requirements.txt
blinker==1.8.2
boto3==1.20.54
botocore==1.23.54
click==8.1.7
colorama==0.4.3
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
MarkupSafe==2.1.5
numpy==2.1.2
opencv-python-headless==4.10.0.84
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.19
Werkzeug==3.0.3
